moved the app specific code to the index.js instead of server.js
